Cleanup, add line item examples
