+++ pkg/edi/segment/l1.go
@@ -8,11 +8,11 @@ import (
 
 // L1 represents the L1 EDI segment
 type L1 struct {
-	LadingLineItemNumber int
-	FreightRate          float64
-	RateValueQualifier   string
-	Charge               float64
-	Percent              float64
+	LadingLineItemNumber     int
+	FreightRate              float64
+	RateValueQualifier       string
+	Charge                   float64
+	SpecialChargeDescription string
 }
 
 // String converts L1 to its X12 single line string representation
@@ -30,21 +30,14 @@ func (s *L1) String(delimiter string) string {
 		"",
 		"",
 		"",
-		"",
-		"",
-		"",
-		"",
-		"",
-		"",
-		"",
-		strconv.FormatFloat(s.Percent, 'f', 4, 64),
+		s.SpecialChargeDescription,
 	}
 	return strings.Join(elements, delimiter) + "\n"
 }
 
 // Parse parses an X12 string that's split into an array into the L1 struct
 func (s *L1) Parse(elements []string) error {
-	expectedNumElements := 19
+	expectedNumElements := 12
 	if len(elements) != expectedNumElements {
 		return fmt.Errorf("L1: Wrong number of elements, expected %d, got %d", expectedNumElements, len(elements))
 	}
@@ -63,6 +56,6 @@ func (s *L1) Parse(elements []string) error {
 	if err != nil {
 		return err
 	}
-	s.Percent, err = strconv.ParseFloat(elements[18], 64)
+	s.SpecialChargeDescription = elements[11]
 	return err
 }
